# perf: prover small fields improvments