--- a/prover/protocol/compiler/globalcs/quotient.go
+++ b/prover/protocol/compiler/globalcs/quotient.go
@@ -5,17 +5,16 @@ import (
 	"reflect"
 	"runtime"
 	"sync"
-	"unsafe"
 
 	"github.com/consensys/linea-monorepo/prover/protocol/coin"
 	"github.com/consensys/linea-monorepo/prover/protocol/variables"
-	"golang.org/x/sync/singleflight"
 
 	"github.com/sirupsen/logrus"
 
 	"github.com/consensys/gnark-crypto/ecc"
 	"github.com/consensys/gnark-crypto/field/koalabear/extensions"
 	"github.com/consensys/gnark-crypto/field/koalabear/fft"
+	"github.com/consensys/linea-monorepo/prover/maths/common/fastpoly"
 	"github.com/consensys/linea-monorepo/prover/maths/common/fastpolyext"
 	"github.com/consensys/linea-monorepo/prover/maths/common/smartvectors"
 	sv "github.com/consensys/linea-monorepo/prover/maths/common/smartvectors"
@@ -27,9 +26,9 @@ import (
 	"github.com/consensys/linea-monorepo/prover/protocol/wizardutils"
 	"github.com/consensys/linea-monorepo/prover/symbolic"
 	"github.com/consensys/linea-monorepo/prover/utils"
+	"github.com/consensys/linea-monorepo/prover/utils/arena"
 	"github.com/consensys/linea-monorepo/prover/utils/collection"
 	"github.com/consensys/linea-monorepo/prover/utils/parallel"
-	ppool "github.com/consensys/linea-monorepo/prover/utils/parallel/pool"
 )
 
 // QuotientCtx collects all the internal fields needed to compute the quotient
@@ -43,9 +42,9 @@ type QuotientCtx struct {
 	// See [mergingCtx.Ratios]
 	Ratios []int
 
-	// ColumnsForRatio[k] stores all the columns involved in the aggregate
+	// ShiftedColumnsForRatio[k] stores all the columns involved in the aggregate
 	// expressions for ratio Ratios[k]
-	ColumnsForRatio [][]ifaces.Column
+	ShiftedColumnsForRatio [][]ifaces.Column
 
 	// RootsPerRatio[k] stores all the root columns involved in the aggregate
 	// expressions for ration Ratios[k]. By root column we mean the underlying
@@ -69,10 +68,6 @@ type QuotientCtx struct {
 	// QuotientShares[k] stores for each k, the list of the Ratios[k] shares
 	// of the quotient for the AggregateExpression[k]
 	QuotientShares [][]ifaces.Column
-
-	// MaxNbExprNode stores the largest number of node AggregateExpressionBoard[*]
-	// has. This is used to dimension the memory pool during the prover time.
-	MaxNbExprNode int
 }
 
 // createQuotientCtx constructs a [quotientCtx] from a list of ratios and aggregated
@@ -81,7 +76,7 @@ type QuotientCtx struct {
 func createQuotientCtx(comp *wizard.CompiledIOP, ratios []int, aggregateExpressions []*symbolic.Expression) QuotientCtx {
 
 	var (
-		allInvolvedHandlesIndex = map[ifaces.ColID]int{}
+		allInvolvedHandlesIndex = map[ifaces.ColID]struct{}{}
 		allInvolvedRootsSet     = collection.NewSet[ifaces.ColID]()
 		_, _, domainSize        = wizardutils.AsExpr(aggregateExpressions[0])
 		ctx                     = QuotientCtx{
@@ -91,7 +86,7 @@ func createQuotientCtx(comp *wizard.CompiledIOP, ratios []int, aggregateExpressi
 			AggregateExpressionsBoard: make([]symbolic.ExpressionBoard, len(ratios)),
 			AllInvolvedColumns:        []ifaces.Column{},
 			AllInvolvedRoots:          []ifaces.Column{},
-			ColumnsForRatio:           make([][]ifaces.Column, len(ratios)),
+			ShiftedColumnsForRatio:    make([][]ifaces.Column, len(ratios)),
 			RootsForRatio:             make([][]ifaces.Column, len(ratios)),
 			QuotientShares:            generateQuotientShares(comp, ratios, domainSize),
 		}
@@ -105,7 +100,6 @@ func createQuotientCtx(comp *wizard.CompiledIOP, ratios []int, aggregateExpressi
 		)
 
 		ctx.AggregateExpressionsBoard[k] = board
-		ctx.MaxNbExprNode = max(ctx.MaxNbExprNode, board.CountNodes())
 
 		// This loop scans the metadata looking for columns with the goal of
 		// populating the collections composing quotientCtx.
@@ -125,10 +119,14 @@ func createQuotientCtx(comp *wizard.CompiledIOP, ratios []int, aggregateExpressi
 			// within a constraint). This works because the symbolic package has
 			// automatic simplification routines that ensure that an expression
 			// does not refer to duplicates of the same variable.
-			ctx.ColumnsForRatio[k] = append(ctx.ColumnsForRatio[k], col)
+			if _, isShifted := col.(column.Shifted); isShifted {
+				// TODO @gbotrel confirm we only get shifted and natural columns.
+				ctx.ShiftedColumnsForRatio[k] = append(ctx.ShiftedColumnsForRatio[k], col)
+			}
 
 			if !uniqueRootsForRatio.Exists(rootCol.GetColID()) {
 				ctx.RootsForRatio[k] = append(ctx.RootsForRatio[k], rootCol)
+				uniqueRootsForRatio.Insert(rootCol.GetColID())
 			}
 
 			// Get the name of the
@@ -136,7 +134,7 @@ func createQuotientCtx(comp *wizard.CompiledIOP, ratios []int, aggregateExpressi
 				continue
 			}
 
-			allInvolvedHandlesIndex[col.GetColID()] = len(ctx.AllInvolvedColumns)
+			allInvolvedHandlesIndex[col.GetColID()] = struct{}{}
 			ctx.AllInvolvedColumns = append(ctx.AllInvolvedColumns, col)
 
 			// If the handle is simply a shift or a natural columns tracks its root
@@ -146,6 +144,9 @@ func createQuotientCtx(comp *wizard.CompiledIOP, ratios []int, aggregateExpressi
 			}
 		}
 	}
+
+	// TODO @gbotrel this context preparation should compute the exact memory needed for the arenas
+	// in Run and prepare them here.
 
 	return ctx
 }
@@ -181,6 +182,9 @@ func (ctx *QuotientCtx) Run(run *wizard.ProverRuntime) {
 	// Compute once the FFT of the natural columns
 
 	domain0 := fft.NewDomain(uint64(ctx.DomainSize), fft.WithCache())
+
+	arenaBase := arena.NewVectorArena[field.Element](ctx.DomainSize * len(ctx.AllInvolvedRoots))
+
 	parallel.Execute(len(ctx.AllInvolvedRoots), func(start, stop int) {
 
 		for k := start; k < stop; k++ {
@@ -198,7 +202,7 @@ func (ctx *QuotientCtx) Run(run *wizard.ProverRuntime) {
 			}
 
 			if smartvectors.IsBase(witness) {
-				res := make([]field.Element, witness.Len())
+				res := arena.Get[field.Element](arenaBase, ctx.DomainSize)
 				witness.WriteInSlice(res)
 				domain0.FFTInverse(res, fft.DIF, fft.WithNbTasks(1))
 				coeffs.Store(name, smartvectors.NewRegular(res))
@@ -226,17 +230,17 @@ func (ctx *QuotientCtx) Run(run *wizard.ProverRuntime) {
 		The first value is ignored because it correspond to the case where w^N = 1
 		(i.e. w is in the smaller subgroup)
 	*/
-	annulatorInvVals := fastpolyext.EvalXnMinusOneOnACoset(ctx.DomainSize, ctx.DomainSize*maxRatio)
-	annulatorInvVals = fext.ParBatchInvert(annulatorInvVals, runtime.GOMAXPROCS(0))
+	var onceAnnulatorExt, onceAnnulatorBase sync.Once
+	var annulatorInvValsExt []fext.Element
+	var annulatorInvVals []field.Element
 
-	scratch := utils.NewScratch[fext.Element](ctx.DomainSize, len(ctx.AllInvolvedRoots))
+	arenaExt := arena.NewVectorArena[fext.Element](ctx.DomainSize * len(ctx.AllInvolvedRoots))
 
 	for i := 0; i < maxRatio; i++ {
 		// reset the scratch offset for this round
-		scratch.Reset()
+		arenaExt.Reset(0)
 
 		// use sync map to store the coset evaluated polynomials
-		var group singleflight.Group
 		computedReeval := sync.Map{} // (ifaces.ColID <=> sv.SmartVector)
 
 		for j, ratio := range ctx.Ratios {
@@ -251,9 +255,9 @@ func (ctx *QuotientCtx) Run(run *wizard.ProverRuntime) {
 			// With the above example, if we are in the ratio = 2 and maxRatio = 8
 			// and i = 1 (it can only be 0 <= i < ratio).
 			var (
-				share   = i * ratio / maxRatio
-				handles = ctx.ColumnsForRatio[j]
-				// roots     = ctx.RootsForRatio[j]
+				share     = i * ratio / maxRatio
+				handles   = ctx.ShiftedColumnsForRatio[j]
+				roots     = ctx.RootsForRatio[j]
 				board     = ctx.AggregateExpressionsBoard[j]
 				metadatas = board.ListVariableMetadata()
 			)
@@ -261,74 +265,57 @@ func (ctx *QuotientCtx) Run(run *wizard.ProverRuntime) {
 			shift := computeShift(uint64(ctx.DomainSize), ratio, share)
 			domain := fft.NewDomain(uint64(ctx.DomainSize), fft.WithShift(shift), fft.WithCache())
 
-			ppool.ExecutePoolChunky(len(handles), func(k int) {
+			parallel.Execute(len(roots), func(start, stop int) {
+				for k := start; k < stop; k++ {
+					root := roots[k]
+					rootName := root.GetColID()
 
-				pol := handles[k]
-				// short-path, the column is a purely Shifted(Natural) or a Natural
-				// (this excludes repeats and/or interleaved columns)
-				root := column.RootParents(pol)
-				rootName := root.GetColID()
-
-				reevaledRoot, _, _ := group.Do(string(rootName), func() (interface{}, error) {
 					// load the coeff
 					_v, _ := coeffs.Load(rootName)
 					vCoeffs := _v.(sv.SmartVector)
 
-					// use a scratch space if possible
-					reevaledRoot := scratch.NewVector()
-
 					// most of the coeffs are regular, so we optimize for that case
 					// and do a fft on the base field.
 					if vr, ok := vCoeffs.(*sv.Regular); ok {
-						vBase := field.Vector(unsafe.Slice((*field.Element)(unsafe.Pointer(&reevaledRoot[0])), ctx.DomainSize))
-						copy(vBase, *vr)
-						domain.FFT(vBase, fft.DIT, fft.OnCoset(), fft.WithNbTasks(1))
+						reevaledRoot := arena.Get[field.Element](arenaExt, ctx.DomainSize)
+						copy(reevaledRoot, *vr)
+						domain.FFT(reevaledRoot, fft.DIT, fft.OnCoset(), fft.WithNbTasks(1))
 
-						// since vBase and reevaledRoot share the same memory, we need to
-						// copy the values backwards
-						// to avoid overwriting values we still need to read.
-						for i := ctx.DomainSize - 1; i >= 0; i-- {
-							fext.SetFromBase(&reevaledRoot[i], &vBase[i])
-						}
-					} else {
-						vCoeffs.WriteInSliceExt(reevaledRoot)
-						domain.FFTExt(reevaledRoot, fft.DIT, fft.OnCoset(), fft.WithNbTasks(1))
+						res := smartvectors.NewRegular(reevaledRoot)
+						computedReeval.Store(rootName, res)
+						continue
 					}
+
+					reevaledRoot := arena.Get[fext.Element](arenaExt, ctx.DomainSize)
+					vCoeffs.WriteInSliceExt(reevaledRoot)
+					domain.FFTExt(reevaledRoot, fft.DIT, fft.OnCoset(), fft.WithNbTasks(1))
 
 					res := smartvectors.NewRegularExt(reevaledRoot)
 					computedReeval.Store(rootName, res)
-
-					return res, nil
-				})
-
-				// Now, we can reuse a soft-rotation of the smart-vector to save memory
-				if !pol.IsComposite() {
-					// in this case, the right vector was the root so we are done
-					return
 				}
+			})
+
+			for _, pol := range handles {
+				root := column.RootParents(pol)
+				rootName := root.GetColID()
+
+				reevaledRoot, _ := computedReeval.Load(rootName)
 
 				if shifted, isShifted := pol.(column.Shifted); isShifted {
 					polName := pol.GetColID()
-					res := sv.SoftRotateExt(reevaledRoot.(sv.SmartVector), shifted.Offset)
+					var res sv.SmartVector
+					switch ssv := reevaledRoot.(sv.SmartVector).(type) {
+					case *sv.Regular:
+						res = sv.SoftRotate(ssv, shifted.Offset)
+					case *sv.RegularExt:
+						res = sv.SoftRotateExt(ssv, shifted.Offset)
+					}
 					computedReeval.Store(polName, res)
-					return
+					continue
 				}
-
-				polName := pol.GetColID()
-				group.Do(string(polName), func() (interface{}, error) {
-					v, _ := coeffs.Load(polName)
-
-					n := uint64(v.(sv.SmartVector).Len())
-
-					res := make([]field.Element, n)
-					v.(sv.SmartVector).WriteInSlice(res)
-					domain.FFT(res, fft.DIT, fft.OnCoset(), fft.WithNbTasks(1))
-
-					// res := sv.FFT(v.(sv.SmartVector), fft.DIT, false, ratio, share, fft.WithNbTasks(2))
-					computedReeval.Store(polName, smartvectors.NewRegular(res))
-					return nil, nil
-				})
-			})
+				panic("never called")
+				// TODO @gbotrel confirm we only get shifted and natural columns.
+			}
 
 			// Evaluates the constraint expression on the coset
 			evalInputs := make([]sv.SmartVector, len(metadatas))
@@ -366,21 +353,33 @@ func (ctx *QuotientCtx) Run(run *wizard.ProverRuntime) {
 
 			// Note that this will panic if the expression contains "no commitment"
 			// This should be caught already by the constructor of the constraint.
-			quotientShare := ctx.AggregateExpressionsBoard[j].Evaluate(evalInputs)
-			if re, ok := quotientShare.(*sv.RegularExt); ok {
-				vq := extensions.Vector(*re)
+			quotientShare := board.Evaluate(evalInputs)
+			switch quotientShare := quotientShare.(type) {
+			case *sv.Regular:
+				onceAnnulatorBase.Do(func() {
+					annulatorInvVals = fastpoly.EvalXnMinusOneOnACoset(ctx.DomainSize, ctx.DomainSize*maxRatio)
+					annulatorInvVals = field.ParBatchInvert(annulatorInvVals, runtime.GOMAXPROCS(0))
+				})
+
+				vq := field.Vector(*quotientShare)
 				vq.ScalarMul(vq, &annulatorInvVals[i])
-			} else {
-				quotientShare = sv.ScalarMulExt(quotientShare, annulatorInvVals[i])
+			case *sv.RegularExt:
+				onceAnnulatorExt.Do(func() {
+					annulatorInvValsExt = fastpolyext.EvalXnMinusOneOnACoset(ctx.DomainSize, ctx.DomainSize*maxRatio)
+					annulatorInvValsExt = fext.ParBatchInvert(annulatorInvValsExt, runtime.GOMAXPROCS(0))
+				})
+				vq := extensions.Vector(*quotientShare)
+				vq.ScalarMul(vq, &annulatorInvValsExt[i])
+			default:
+				// quotientShare = sv.ScalarMulExt(quotientShare, annulatorInvVals[i])
+				utils.Panic("unexpected type %T", quotientShare)
 			}
 
 			run.AssignColumn(ctx.QuotientShares[j][share].GetColID(), quotientShare)
-
 		}
 
 		// Forcefully clean the memory for the computed reevals
 		computedReeval = sync.Map{}
-		group = singleflight.Group{}
 	}
 
 	logrus.Infof("[global-constraint] msg=\"computed the quotient\"")