--- a/prover/protocol/compiler/logderivativesum/prover_tasks.go
+++ b/prover/protocol/compiler/logderivativesum/prover_tasks.go
@@ -207,6 +207,7 @@ func (a MAssignmentTask) Run(run *wizard.ProverRuntime) {
 
 		for frag := range a.T {
 			tCollapsed[frag] = wizardutils.RandLinCombColAssignment(run, collapsingRandomness, a.T[frag])
+			fragmentUnionSize += tCollapsed[frag].Len()
 		}
 
 		for i := range a.S {
@@ -225,7 +226,7 @@ func (a MAssignmentTask) Run(run *wizard.ProverRuntime) {
 		//
 		// It is used to let us know where an entry of S appears in T. The stored
 		// 2-uple of integers indicate [fragment, row]
-		mapM = make(map[fext.Element][2]int, fragmentUnionSize)
+		mapM = make(map[fext.Element][2]uint32, fragmentUnionSize)
 	)
 
 	// This loops initializes mapM so that it tracks to the positions of the
@@ -252,9 +253,11 @@ func (a MAssignmentTask) Run(run *wizard.ProverRuntime) {
 
 		m[frag] = make([]field.Element, tCollapsed[frag].Len())
 
-		for k := max(0, start); k < min(size, end); k++ {
+		start = max(0, start)
+		end = min(size, end)
+		for k := start; k < end; k++ {
 			v := tCollapsed[frag].GetExt(k)
-			mapM[v] = [2]int{frag, k}
+			mapM[v] = [2]uint32{uint32(frag), uint32(k)}
 		}
 	}