… prover/protocol/compiler/globalcs/factoring.go
… prover/protocol/compiler/mpts/prover.go
… prover/protocol/dedicated/expr_handle/expr_handle.go
… prover/protocol/dedicated/expr_handle/rand_lin_comp_expr.go
… prover/protocol/serialization/serializer.go
… prover/symbolic/constant.go
… prover/symbolic/constructors.go
… prover/symbolic/csv.go
… prover/symbolic/evaluation.go
… prover/symbolic/expression.go
… prover/symbolic/expression_board.go
… prover/symbolic/lincomb.go
… prover/symbolic/pattern.go
… prover/symbolic/polyeval.go
… prover/symbolic/product.go
… prover/symbolic/simplification.go
… prover/symbolic/simplify/cost_stat.go
… prover/symbolic/simplify/cost_stat_test.go
… prover/symbolic/simplify/factor.go
… prover/symbolic/simplify/factor_test.go
… prover/symbolic/simplify/rmpolyeval.go
… prover/symbolic/simplify/simplify.go
… prover/symbolic/variable.go
… prover/utils/arena/vector_arena.go
… prover/utils/parallel/parallel.go
… prover/utils/scratch.go