This PR addresses couple of perf issues, will let copilot generated the summary, but essentially: use the vector arena to allocate large blocks of memory in the evaluation board and the quotient compute (contiguous memory is good for these compute heavy func), hunt down some unecessary mem allocs or ops on the extension instead of the base field. 
Remove factorization of symbolic expression before quotient for now, may re-introduce later if worth it (unsure).

ESH in symbolic is now a field ext directly.

parallel.Execute have a small heuristic to ensure we don't create too many go routines.

smart vectors need a lot of clean up, will come in a future PR.

## Benchmark

```
benchmark                                  old ns/op      new ns/op      delta
BenchmarkCompiler/realistic-segment-96     7571576095     5494496541     -27.43%

benchmark                                  old allocs     new allocs     delta
BenchmarkCompiler/realistic-segment-96     94413422       93385987       -1.09%

benchmark                                  old bytes       new bytes       delta
BenchmarkCompiler/realistic-segment-96     63911328248     44511278496     -30.35%
```